requirements.txt
streamlit==1.20.0
pandas==1.5.3
numpy==1.24.2
plotly==5.13.1
gspread==5.7.0
google-auth
